changing controller layer name from resource to controller
