+++ src/main/java/com/syst/trades/controller/ClientTradeResource.java
@@ -1,4 +1,4 @@
-package com.syst.trades.resource;
+package com.syst.trades.controller;
 
 import java.net.URI;
 import java.util.List;
@@ -9,34 +9,34 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import com.syst.trades.model.HistoryProductQuantity;
-import com.syst.trades.repository.HistoryProductQuantityRepository;
+import com.syst.trades.model.ClientTradeDebit;
+import com.syst.trades.repository.ClientTradeDebitRepository;
 
 @RestController
-@RequestMapping("/history-product-quantity")
-public class HistoryProductQuantityResource {
+@RequestMapping("/client-trade")
+public class ClientTradeResource {
 
 	@Autowired
-	private HistoryProductQuantityRepository historyProductQuantityRepository;
+	private ClientTradeDebitRepository clientTradeDebitRepository;
 
 	@GetMapping
-	public List<HistoryProductQuantity> toList() {
-		return historyProductQuantityRepository.findAll();
+	public List<ClientTradeDebit> toList() {
+		return clientTradeDebitRepository.findAll();
 	}
 
 	@PostMapping
 	@ResponseStatus(HttpStatus.CREATED)
-	public void toSave(HistoryProductQuantity historyProductQuantity, HttpServletResponse response) {
-		HistoryProductQuantity historyProductQuantitySaved =
-				historyProductQuantityRepository.save(historyProductQuantity);
+	public void toSave(@RequestBody ClientTradeDebit clientTradeDebit, HttpServletResponse response) {
+		ClientTradeDebit clientTradeDebitSaved = clientTradeDebitRepository.save(clientTradeDebit);
 
 		URI uri = ServletUriComponentsBuilder.fromCurrentRequestUri().path("/{id}")
-				.buildAndExpand(historyProductQuantitySaved.getId()).toUri();
+				.buildAndExpand(clientTradeDebitSaved.getId()).toUri();
 
 		response.setHeader("Location", uri.toASCIIString());
 	}
