+++ src/main/java/com/syst/trades/controller/HistoryProductQuantityResource.java
@@ -1,4 +1,4 @@
-package com.syst.trades.resource;
+package com.syst.trades.controller;
 
 import java.net.URI;
 import java.util.List;
@@ -9,34 +9,34 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import com.syst.trades.model.Status;
-import com.syst.trades.repository.StatusRepository;
+import com.syst.trades.model.HistoryProductQuantity;
+import com.syst.trades.repository.HistoryProductQuantityRepository;
 
 @RestController
-@RequestMapping("/status")
-public class StatusResource {
+@RequestMapping("/history-product-quantity")
+public class HistoryProductQuantityResource {
 
 	@Autowired
-	private StatusRepository statusRepository;
+	private HistoryProductQuantityRepository historyProductQuantityRepository;
 
 	@GetMapping
-	public List<Status> toList() {
-		return statusRepository.findAll();
+	public List<HistoryProductQuantity> toList() {
+		return historyProductQuantityRepository.findAll();
 	}
 
 	@PostMapping
 	@ResponseStatus(HttpStatus.CREATED)
-	public void toSave(@RequestBody Status status, HttpServletResponse response) {
-		Status statusSaved = statusRepository.save(status);
+	public void toSave(HistoryProductQuantity historyProductQuantity, HttpServletResponse response) {
+		HistoryProductQuantity historyProductQuantitySaved =
+				historyProductQuantityRepository.save(historyProductQuantity);
 
 		URI uri = ServletUriComponentsBuilder.fromCurrentRequestUri().path("/{id}")
-				.buildAndExpand(statusSaved.getId()).toUri();
+				.buildAndExpand(historyProductQuantitySaved.getId()).toUri();
 
 		response.setHeader("Location", uri.toASCIIString());
 	}
